Added point multipliers to .guess or .bet commands that guess correctly in the first 4 tries

+++ StormBot/Database/Entities/StormPlayerDataEntity.cs
@@ -10,6 +10,6 @@ namespace StormBot.Database.Entities
 
 		public ulong DiscordID { get; set; }
 
-		public int Wallet { get; set; }
+		public double Wallet { get; set; }
 	}
 }
